requirements.txt
flask
line-bot-sdk
pytest-playwright
pillow
pytz